modulo detalhes dos planos completo

+++ app/Http/Controllers/Admin/PlanController.php
@@ -52,11 +52,21 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+                        ->with('details')
+                        ->where('url', $url)
+                        ->first();
 
         if(!$plan)
             return redirect()->back();
 
+        if($plan->details->count() > 0){
+
+            return redirect()->back()
+                             ->with('error', 'registro não pode ser excluído, possue dependências');
+        }
+        
+      
         $plan->delete();
 
         return redirect()->route('plans.index');
